requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0,<3.0
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
openpyxl
python-pptx
python-dotenv
google-generativeai
pandas
numpy
requests